# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Billing application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  costPrice: number;
  marginPercent: number;
  sellingPrice: number;
  gstPercent: number;
  stock: number;
  hsn?: string;
  unit?: string;
  createdAt: string;
}

export interface CreateProductBody {
  name: string;
  costPrice: number;
  marginPercent: number;
  gstPercent: number;
  stock: number;
  hsn?: string;
  unit?: string;
}

export interface UpdateProductBody {
  name?: string;
  costPrice?: number;
  marginPercent?: number;
  gstPercent?: number;
  stock?: number;
  hsn?: string;
  unit?: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface Customer {
  id: number;
  name: string;
  phone?: string;
  email?: string;
  address?: string;
  gstin?: string;
  totalPurchases?: number;
  createdAt: string;
}

export interface InvoiceSummary {
  id: number;
  invoiceNumber: string;
  customerName?: string;
  date: string;
  grandTotal: number;
  totalProfit?: number;
  itemCount?: number;
  createdAt: string;
}

export interface CustomerDetail {
  id: number;
  name: string;
  phone?: string;
  email?: string;
  address?: string;
  gstin?: string;
  totalPurchases?: number;
  createdAt: string;
  invoices: InvoiceSummary[];
}

export interface CreateCustomerBody {
  name: string;
  phone?: string;
  email?: string;
  address?: string;
  gstin?: string;
}

export interface UpdateCustomerBody {
  name?: string;
  phone?: string;
  email?: string;
  address?: string;
  gstin?: string;
}

export interface CustomerListResponse {
  customers: Customer[];
  total: number;
  page: number;
  limit: number;
}

export interface InvoiceItem {
  id?: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  costPrice?: number;
  gstPercent: number;
  gstAmount: number;
  discountPercent?: number;
  discountAmount?: number;
  totalAmount: number;
  profit?: number;
}

export interface CreateInvoiceItem {
  productId: number;
  quantity: number;
  unitPrice: number;
  discountPercent?: number;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  customerId?: number;
  customerName?: string;
  date: string;
  items: InvoiceItem[];
  subtotal: number;
  totalGst: number;
  totalDiscount: number;
  grandTotal: number;
  totalProfit?: number;
  notes?: string;
  createdAt: string;
}

export interface CreateInvoiceBody {
  customerId?: number;
  date: string;
  items: CreateInvoiceItem[];
  notes?: string;
  overallDiscountPercent?: number;
}

export interface InvoiceListResponse {
  invoices: InvoiceSummary[];
  total: number;
  page: number;
  limit: number;
}

export interface DashboardSummary {
  totalSales: number;
  totalProfit: number;
  totalGst?: number;
  totalInvoices: number;
  totalCustomers: number;
  totalProducts: number;
  financialYear?: string;
}

export interface TopProduct {
  productId: number;
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
  totalProfit?: number;
}

export interface TopCustomer {
  customerId: number;
  customerName: string;
  totalPurchases: number;
  invoiceCount: number;
}

export interface MonthlySales {
  month: number;
  year: number;
  monthName?: string;
  totalSales: number;
  totalProfit: number;
  totalGst?: number;
  invoiceCount: number;
}

export interface SalesReportEntry {
  period: string;
  sales: number;
  gst: number;
  profit?: number;
  invoiceCount: number;
}

export interface SalesReport {
  entries: SalesReportEntry[];
  totalSales: number;
  totalGst: number;
  totalProfit?: number;
}

export interface GstReportEntry {
  month: number;
  monthName: string;
  cgst: number;
  sgst: number;
  igst: number;
  totalGst: number;
  taxableAmount?: number;
}

export interface GstReport {
  entries: GstReportEntry[];
  totalCgst: number;
  totalSgst: number;
  totalIgst: number;
  totalGst: number;
  financialYear?: string;
}

export interface ProfitReportEntry {
  period: string;
  revenue: number;
  cost: number;
  profit: number;
  margin?: number;
}

export interface ProfitReport {
  entries: ProfitReportEntry[];
  totalRevenue: number;
  totalCost: number;
  totalProfit: number;
  profitMargin?: number;
}

export interface ProductWiseReport {
  productId: number;
  productName: string;
  totalQuantity: number;
  totalRevenue: number;
  totalCost?: number;
  totalProfit: number;
  totalGst?: number;
}

export interface CustomerWiseReport {
  customerId: number;
  customerName: string;
  totalPurchases: number;
  invoiceCount: number;
  lastPurchaseDate?: string;
}

export interface FinancialYear {
  label: string;
  startDate: string;
  endDate: string;
}

export interface DatabaseExport {
  version: string;
  exportedAt: string;
  products: Product[];
  customers: Customer[];
  invoices: Invoice[];
  invoiceItems: InvoiceItem[];
}

export type ImportResultImported = {
  products?: number;
  customers?: number;
  invoices?: number;
};

export interface ImportResult {
  success: boolean;
  message: string;
  imported?: ImportResultImported;
}

export interface DeleteResponse {
  success: boolean;
  message?: string;
}

export type ListProductsParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type ListCustomersParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type ListInvoicesParams = {
  customerId?: number;
  startDate?: string;
  endDate?: string;
  page?: number;
  limit?: number;
};

export type GetDashboardSummaryParams = {
  /**
   * Financial year e.g. 2025-2026
   */
  financialYear?: string;
};

export type GetRecentTransactionsParams = {
  limit?: number;
};

export type GetTopProductsParams = {
  financialYear?: string;
  limit?: number;
};

export type GetTopCustomersParams = {
  financialYear?: string;
  limit?: number;
};

export type GetMonthlySalesParams = {
  financialYear?: string;
};

export type GetSalesReportParams = {
  startDate: string;
  endDate: string;
  groupBy?: GetSalesReportGroupBy;
};

export type GetSalesReportGroupBy =
  (typeof GetSalesReportGroupBy)[keyof typeof GetSalesReportGroupBy];

export const GetSalesReportGroupBy = {
  daily: "daily",
  monthly: "monthly",
  yearly: "yearly",
} as const;

export type GetGstReportParams = {
  financialYear: string;
};

export type GetProfitReportParams = {
  startDate: string;
  endDate: string;
  groupBy?: GetProfitReportGroupBy;
};

export type GetProfitReportGroupBy =
  (typeof GetProfitReportGroupBy)[keyof typeof GetProfitReportGroupBy];

export const GetProfitReportGroupBy = {
  daily: "daily",
  monthly: "monthly",
  yearly: "yearly",
} as const;

export type GetProductWiseReportParams = {
  startDate: string;
  endDate: string;
};

export type GetCustomerWiseReportParams = {
  startDate: string;
  endDate: string;
};