# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Billing application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /customers/{id}/purchases:
    get:
      operationId: getCustomerPurchases
      tags: [customers]
      summary: Get purchase history for a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceSummary"

  # ── Invoices ──
...
  /backup/import:
    post:
      operationId: importDatabase
      tags: [backup]
      summary: Import database from JSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatabaseExport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # ── Financial Years ──
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, costPrice, marginPercent, sellingPrice, gstPercent, stock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        costPrice:
          type: number
        marginPercent:
          type: number
        sellingPrice:
          type: number
        gstPercent:
          type: number
        stock:
          type: integer
        hsn:
          type: string
        unit:
          type: string
        createdAt:
          type: string
...
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: string
        totalPurchases:
          type: number
        createdAt:
          type: string
    CustomerDetail:
      type: object
      required: [id, name, createdAt, invoices]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: string
        totalPurchases:
          type: number
        createdAt:
          type: string
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceSummary"
    CreateCustomerBody:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: string
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        gstin:
          type: string
    CustomerListResponse:
      type: object
      required: [customers, total, page, limit]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    InvoiceItem:
      type: object
      required: [productId, productName, quantity, unitPrice, gstPercent, gstAmount, totalAmount]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        costPrice:
          type: number
        gstPercent:
          type: number
        gstAmount:
          type: number
        discountPercent:
          type: number
        discountAmount:
          type: number
        totalAmount:
          type: number
        profit:
          type: number
...
    Invoice:
      type: object
      required: [id, invoiceNumber, date, items, subtotal, totalGst, totalDiscount, grandTotal, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        date:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        subtotal:
          type: number
        totalGst:
          type: number
        totalDiscount:
          type: number
        grandTotal:
          type: number
        totalProfit:
          type: number
        notes:
          type: string
        createdAt:
          type: string
...
    InvoiceSummary:
      type: object
      required: [id, invoiceNumber, date, grandTotal, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerName:
          type: string
        date:
          type: string
        grandTotal:
          type: number
        totalProfit:
          type: number
        itemCount:
          type: integer
        createdAt:
          type: string
...
    DatabaseExport:
      type: object
      required: [version, exportedAt, products, customers, invoices, invoiceItems]
      properties:
        version:
          type: string
        exportedAt:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        invoiceItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    ImportResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        imported:
          type: object
          properties:
            products:
              type: integer
            customers:
              type: integer
            invoices:
              type: integer
    DeleteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string