# lib/api-zod/src/generated/types/databaseExport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Billing application API
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from "./customer";
import type { Invoice } from "./invoice";
import type { InvoiceItem } from "./invoiceItem";
import type { Product } from "./product";

export interface DatabaseExport {
  version: string;
  exportedAt: string;
  products: Product[];
  customers: Customer[];
  invoices: Invoice[];
  invoiceItems: InvoiceItem[];
}