# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Billing application API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  costPrice: number;
  marginPercent: number;
  sellingPrice: number;
  gstPercent: number;
  stock: number;
  hsn?: string;
  unit?: string;
  createdAt: string;
}